• use #!/bin/sh as shebang for “Ruby → Run” and “TextMate → Relaunch”.
This is to prevent running ‘Support/lib/bash_init.sh’ which sets up RUBYLIB. Having RUBYLIB set to include Support/lib means users who run scripts that require libraries named the same as those we have in TextMate will likely get the wrong one.

Very few commands actually make use of the fact that RUBYLIB is set, so maybe we should edit ‘bash_init.sh’ to simply not set it, although I am waiting with such change till after next build, so we have time to thoroughly test that everything still works.

The downside by this change is that running a ruby script in TM intended to be used by a TM command, now require TM support libraries to be included using the full path (or the Support/lib folder be added to the include path).

On that note, RubyMate/stdin_dialog.rb had to be updated to do so, since it is indirectly loaded when running scripts (now in an environment w/o Support/lib in RUBYLIB).


git-svn-id: http://macromates.com/svn/Bundles/trunk@8572 dfb7d73b-c2ec-0310-8fea-fb051d288c6d