Adding GroovyMate (which shows off tm_dialog_read).
On 10.4, if the running Groovy script requests input on STDIN and none is available, a dialog is presented to the user to capture the input. This currently doesn't work on 10.5, it will behave "normally" by sitting and waiting for input. This is a current limitation to tm_dialog_read.dylib in that we don't know how to make it work on Leopard.

git-svn-id: http://macromates.com/svn/Bundles/trunk@9054 dfb7d73b-c2ec-0310-8fea-fb051d288c6d