• no longer conditionally disabled the OK button when the text field is empty.
This is because in some cases we do actually want to allow the user to pass an empty string.

I am going to add a boolean in the property list to control whether or not the OK button should conditionally be disabled if that is possible (will look into that part later).


git-svn-id: http://macromates.com/svn/Bundles/trunk@9132 dfb7d73b-c2ec-0310-8fea-fb051d288c6d