• Complete rewrite from scratch.
With the previous grammar, it was more or less impossible to properly support anonymous inner classes and runtime created object arrays. To try and support these things and other advanced constructs, we try and scope more context.

This is probably about 95% useable. It supports more than the previous grammar, but may have some bugs.



git-svn-id: http://macromates.com/svn/Bundles/trunk@9374 dfb7d73b-c2ec-0310-8fea-fb051d288c6d