• Provide overridden version of select() in order to indicate that data is actually available on fds that we are using interactive input for (i.e. is stdin, or a dup of, and is owned by TM's pid).
This may be too naive an approach. We are changing the semantics of select() in that if the set of fds it passes for reading contain fds that are eligible for using the dialog to get input for we return the count of such fds, ignoring what select() would have returned. Preliminary tests show this to be not a problem, but if we start seeing weird behaviour this would be a likely candidate for the cause.

git-svn-id: http://macromates.com/svn/Bundles/trunk@9593 dfb7d73b-c2ec-0310-8fea-fb051d288c6d