• when posting to a blog we let the dateCreated header be the time of the local system (w/o time zone info) and additionally send a date_created_gmt header which is the time in GMT. • when receiving data from the blog, we first check for a date_created_gmt header and use that (as the time in GMT) and fall back on dateCreated, which we interpret as in the users local time zone.
The date_created_gmt is a new header introduced by WordPress and is supposed to solve the problem with the missing time zone info in the standard. The (new) semantics of dateCreated (i.e. treating times as in the users local time zone) seems to be the consensus (prior to the GMT header) despite being rather fragile.


git-svn-id: http://macromates.com/svn/Bundles/trunk@9600 dfb7d73b-c2ec-0310-8fea-fb051d288c6d