added the following buttons to Advanced Drawer: - Reveal Target Folder in Finder - Open Bundle Editor
I tested the entire bundle with a standard user (no admin rights). If that user wants to install something into 'App Bundle' or 'Lib Bundle' a file access right error occurs in the Log. The "funny" thing is that if that user presses the button to upgrade TM's Support/lib nothing happens. This is the desired behaviour but I DO NOT KNOW WHY?

git-svn-id: http://macromates.com/svn/Bundles/trunk@10344 dfb7d73b-c2ec-0310-8fea-fb051d288c6d