Disable auto-saving of the current document in the LaTeX typeset and view command.  This was messing with people’s file encodings.
To re-enable the typesetting of unsaved documents, you must change the “Save” setting
of this command to “Current File” and add the variable TM_LATEX_AUTOSAVE to TextMate’s
Shell Variables preferences. Be warned that your document must be encoded as UTF-8 if
you exercise this option — because TextMate.save_current_document cannot know the
file encoding you prefer.



git-svn-id: http://svn.textmate.org/trunk@10991 dfb7d73b-c2ec-0310-8fea-fb051d288c6d