• Support for scoping macros and their parameters.
This is deliberately done in a way where it is possible to “plug in” specific macros to get more specific scoping (e.g. explicitly supporting the ‘code’ macro so the contents are scoped as the correct language).

Also, macros with a body are not scoped as a block. This is impossible to do generally as there is no syntactical indicator as to whether a macro is going to be followed by a body or not.


git-svn-id: http://svn.textmate.org/trunk@11127 dfb7d73b-c2ec-0310-8fea-fb051d288c6d