Change tab triggers: It’s now clm⇥ and catm⇥ for the implementation part of the class/category.
This is more symmetrical with clh⇥ and cath⇥ plus (at least for me) a better metaphor (i.e. I think of it as the h and m file, not the “interface” and “implementation” file, especially considering that the interface snippets did get ‘h’).

Also fix the cath⇥ trigger (which erroneously was cati⇥).


git-svn-id: http://svn.textmate.org/trunk@11217 dfb7d73b-c2ec-0310-8fea-fb051d288c6d