Fix syntax for shell functions
The matched syntax is still more inclusive than what bash actually
supports, but as the full syntax for bash is undocumented, being
more inclusive is better than being less. The main motivation for
this change was to properly ignore = in identifiers so defining
an empty array does not look like a function definition.

git-svn-id: http://svn.textmate.org/trunk@11738 dfb7d73b-c2ec-0310-8fea-fb051d288c6d