Add scripts to build the keychain and plist ruby extensions with 64-bit architectures.
Since this requires multiple steps (building as both 32- and 64-bit and then folding together with lipo) I could not find a way to do this solely with ruby’s mkmf, so instead I’ve created a wrapper script which performs each step in turn.
Prior to this commit the extconf files specified the 10.4u SDK, but due to the 10.4u CoreFoundation framework not containing 64-bit architectures I have removed this flag (so the bundles will be built against the user’s system version).
The plist bundle has been built against the 10.5 SDK for some time now, apparently without issue from users on 10.4, so I think this required change is fine.


git-svn-id: http://svn.textmate.org/trunk@11741 dfb7d73b-c2ec-0310-8fea-fb051d288c6d