Remove TextMate plug-in template.
This plug-in contain class dumps as API headers. I am fine with people doing hacks in their plug-ins given that there is no actual API, but we should *not* host class dumps in the official repository as this is *not* officially a supported way to write your plug-in and it *will* break in the future!


git-svn-id: http://svn.textmate.org/trunk@11919 dfb7d73b-c2ec-0310-8fea-fb051d288c6d