Rework (web) output theme code.
It’s now possible to set the TM_THEME_PATH variable to a colon-separated list of places where themes are stored. The idea is to move themes to bundles and have bundles with themes append to this variable via the shell variable settings.

Theme names shown in the pop-up are capitalized and sorted.

Unified the code dealing with default, bundle, and user styles. I haven’t matched the previous semantics, as it would only use styles from the current bundle if those styles had the same name as one of the default styles (excluding user styles). It now link to all styles matching ${TM_BUNDLE_SUPPORT}/Support/css/*/styles.css. This means a bundle can cause new items to appear in the theme selector (which is likely not desired).



git-svn-id: http://svn.textmate.org/trunk@11983 dfb7d73b-c2ec-0310-8fea-fb051d288c6d