Insert newlines in bundle description.
The bundle description is supposed to be a single line of text without HTML tags (beyond <a href="…">). Programs showing the description (not capable of showing it as HTML) will strip out HTML tags and/or only show to the first newline.

Without these newlines, stripping the HTML tags would give a bad result (missing spacing).


git-svn-id: http://svn.textmate.org/trunk@11995 dfb7d73b-c2ec-0310-8fea-fb051d288c6d